SM: seperated out interfaces and split up some duplicates to shared
