+++ src/modules/challenge-broken-authentication/levels/1/code/interfaces.ts
@@ -1,0 +1,17 @@
+interface Document {
+    name: string,
+    private: boolean,
+    category: string,
+    fileContents: string
+}
+
+interface DocumentCategory {
+    name: string,
+    private: boolean,
+    description: string
+}
+
+export {
+    Document,
+    DocumentCategory
+};
