+++ src/modules/challenge-broken-authentication/levels/2/code/interfaces.ts
@@ -1,0 +1,13 @@
+import { AuthenticateUser } from "../../../../../shared/sharedInterfaces";
+
+interface Document {
+    name: string,
+    private: boolean,
+    category: string,
+    fileContents: string
+}
+
+export {
+    AuthenticateUser,
+    Document
+};
