--- src/modules/challenge-broken-authentication/module-types.ts
@@ -1,10 +1,0 @@
-export interface Document {
-    name: string,
-    private: boolean,
-    category: string,
-    fileContents: string
-}
-
-export type GetDocumentsArguments = {
-    filterByPrivateDocuments: boolean
-}
