+++ src/modules/challenge-custom-scalar-validation/levels/1/code/JSONString.ts
