+++ src/modules/challenge-custom-scalar-validation/levels/1/code/interfaces.ts
@@ -1,0 +1,24 @@
+interface resetPasswordRequestResult {
+    user_id: string,
+    reset_token: string
+}
+
+interface Payload {
+    resetToken: string,
+    userId: number
+}
+
+interface ResetCustomerPasswordArguments {
+    payload: Payload
+}
+
+interface ResetCustomerPasswordResponse {
+    result: string,
+    debug: string
+}
+
+export {
+    ResetCustomerPasswordArguments,
+    ResetCustomerPasswordResponse,
+    resetPasswordRequestResult
+};
