+++ src/modules/challenge-malicious-queries/levels/2/code/interfaces.ts
@@ -1,0 +1,17 @@
+import { Module } from "module";
+import { Product } from "../../../../../shared/sharedInterfaces";
+
+interface ModuleProduct extends Product {
+    categories?: Category[] | number[]
+}
+
+interface Category {
+    categoryId: number,
+    name: string,
+    products: number[] | ModuleProduct[]
+}
+
+export {
+    Category,
+    ModuleProduct
+};
