+++ src/modules/challenge-misconfigured-access-control/levels/1/code/interfaces.ts
@@ -1,0 +1,13 @@
+import { User } from "../../../../../shared/sharedInterfaces";
+
+interface Order {
+    orderId: number,
+    orderNote: string,
+    totalCost: number
+}
+
+interface ModuleUser extends User {
+    orders: Order[]
+}
+
+export { ModuleUser };
