+++ src/modules/challenge-sql-injection/levels/1/code/interfaces.ts
@@ -1,0 +1,20 @@
+import { AuthenticateUser } from "../../../../../shared/sharedInterfaces";
+
+interface AuthenticateAdminResponse {
+    user: {
+        userId: number,
+        email: string
+    }
+    token: string
+}
+
+interface DbUser {
+    user_id: number,
+    email: string
+}
+
+export {
+    AuthenticateUser,
+    AuthenticateAdminResponse,
+    DbUser
+};
